+++ tests/expect/test/index.test.ts
@@ -186,25 +186,4 @@ describe('Expect API', () => {
       logs.find((log) => log.includes('Tests 1 failed | 1 passed')),
     ).toBeTruthy();
   }, 12000);
-
-  it('no unexpected rpc error about result.expected', async () => {
-    const __filename = fileURLToPath(import.meta.url);
-    const __dirname = dirname(__filename);
-
-    const { cli } = await runRstestCli({
-      command: 'rstest',
-      args: ['run', 'fixtures/error.test.ts'],
-      options: {
-        nodeOptions: {
-          cwd: __dirname,
-        },
-      },
-    });
-    await cli.exec;
-    expect(cli.exec.process?.exitCode).toBe(1);
-
-    const logs = cli.stdout.split('\n').filter(Boolean);
-
-    expect(logs.find((log) => log.includes('Error: Symbol('))).toBeFalsy();
-  });
 });
